Let `File.write` and `File.write!` take advantage of `writev`
As a performance optimization, have these functions use `:file.write/3`,
usually with `:raw` mode, instead of `:file.write_file/3`. This will
cause the BEAM to use the `writev` system call instead of `write`.

The advantage of this is that, when writing an iolist, the BEAM need not
concatenate the items and produce a single binary for writing. Instead,
it can let the operating system read each item from memory individually;
the final output is produced only in the target file. Skipping the
binary concatenation saves work, saves memory, and creates less garbage
for later collection.

Note that `:raw` mode cannot be used when a character encoding is given.
In `:raw` mode, the BEAM does not create a separate process to handle
the file, and (I think) it would normally be this process which performs
the character encoding. So if an encoding like `:utf8` is given, we do
not use `:raw` mode and therefore do not use `writev`.