Explain race condition in assert_receive to user when it happens
When using assert_receive, it is possible for the following sequence of
events:
1. `after` timeout triggers, indicating messages was not received
2. Process receive messages
3. `Process.info(self(), :messages) is called and gets message
`assert_receive` was waiting for.

Previously, `assert_receive` would not check the output
`Process.info(self, :messages)` for the expected pattern, so it would
print a confusing error message that said "No message matching <pattern>
after <timeout>ms", but then show a matching message in the "Process
mailbox" section.  Now, for this special race-condition, a long-form
explanation that the message is being delivered at or near the timeout
is being printed, along with how this could impact production and how to
fix it if it only impacts tests.