Use Kernel.to_string/1 for conversion
Kernel.to_string/1 is actually a macro that expands to literally what is
being changed, but it's probably a little cleaner to go ahead and use
that abstraction rather than call to the protocol directly.

It seems this was originally introduced (https://github.com/iamvery/elixir/commit/071b0830f8e81b89e0d43ca9f16d288510559361)
with a case statement to preserve line numbers. That appears to no
longer be the case.