Simplify Kernel.match?/2
There's no need for a special handling of the underscore as the pattern,
the case expression is able of optimizing such cases too.

Additionally the right was never evaluated with underscore as the
pattern, which could lead to subtle bugs.

Currently match?(_, foo) will emit a pattern warning since both branches
of the case are exactly the same. This means such expressions had to be
removed from the test suite. They were replaced with the equivalent:

    fn _ -> true end