Clarify how to use test_stdlib instead of test_elixir
I always thought test_elixir was broken as from the directory
structure I'd have expected it to only run the tests in elixir
e.g. the standard library which is what I mostly touch around
here :) However, it always ran almost everything.

A look at the Makefile reveals that this seems intentional but
that there's also test_stdlib, which seems for to point out in
the README imo.