Give context to invalid pattern error
Talking with @fishcakez we thought that the error message needed a
little more context to help developers find the problem and fix it. This
also addresses the fact that this can happen outside of a `case` match.

~~~elixir
  def fun(true && true), do: :ok
  def fun(other_fun()), do: :ok
~~~

Thanks to @fishcakez for the help and the examples.

Amos King @adkron <amos@binarynoggin.com>