Make test fail if user has a git template without “info/“ (#5525)
Git provides a global template mechanism, to support initializing new
repos with e.g. hook scripts. If you don’t put an “info” subdirectory
in your template, new repos won’t have one, which causes a few of the
sparse-checkout tests to fail, including this one.

This test will cause the failure even when the user has no template
configured, by creating an empty template for the duration of this
test.

(Because it’s using an environment variable, the other sparse-
checkout tests could fail too if they happen to run concurrently with
this one, but the fix for this fixes all those failures.)