Check for "unbound variable _" in elixir_exp (#5773)
I also had to adjust the check for "_ -> ..." clauses in "cond" because right
now, it performs such check after expanding the clauses, but with the change in
this commit then the "unbound variable _" error would fire first. With this
commit, we check for "_ -> ..." clauses before expanding the clauses, which may
not be ideal since the clauses could expand to "_ -> ..." but it's a compromise
we're willing to make (after all, this error is just a friendlier error than
"unbound variable _", not a necessary one).