Fix the behaviour of defdelegate/2 with default arguments
When default arguements (\\) are used with defdelegate/2, a function is
generated for each default argument and each generated function calls the
delegate target function with the same arity. For example:

    defdelegate demonitor(monitor_ref, options \\ []), to: :erlang

will be expanded to

    def demonitor(monitor_ref), do: :erlang.demonitor(monitor_ref)
    def demonitor(monitor_ref, options), do: :erlang.demonitor(monitor_ref, options)

However, this seems to "break" the expectation that optional arguments set: when
a function has optional arguments, all the additional functions with lower arity
that are generated by the default arguments only call the function with maximum
arity with the default values provided for the optional arguments.

This means that if :erlang.demonitor/1 had a different behaviour than
:erlang.demonitor/2, then it would be hard to guess why because the expectation
is that demonitor(monitor_ref) only calls demonitor(monitor_ref, []), ending up
always to :erlang.demonitor/2.

This commit fixes this by leaving the signature of the functions to delegate
untouched (with optional arguments if present).