Add original values to context merging in ExUnit
Closes #5898.

Since 58377cd, `ExUnit.Callbacks.__merge/3` is a recursive function which
drops the `:ok` atom from the passed value if the value is an ok-tuple.
This caused potential errors raised by `raise_merge_failed!/2` to show
only part of the passed value in the error message.

Returning `{:ok, "foo"}` from a setup block produced an error like this:

    ** (RuntimeError) expected [...] got \"foo\" instead

This patch adds `ExUnit.Callbacks.__merge/4`, which is called by
`__merge__/3` with adds the original value as the fourth argument, to be
able to display the full value:

    ** (RuntimeError) expected [...] got {:ok, \"foo\"} instead