Optimize Regex.escape/1
Use a similar optimization technique to the one in plug https://github.com/elixir-lang/plug/pull/547

This allows us to keep as large parts of the initial binary as possible,
in the case of no escapes, no allocation is needed.

In most cases it should be about twice as fast, in case no escape is needed,
in can be even 10x faster. For small strings, it will be about 15% slower, but
with entire execution in range of 50μs, this should be acceptable.