Document manager behaviour, closes #5830
It was decided to not raise on manager conflicts because
it can lead to many false positives. For example, if a
dependency comes from Mix and Rebar 3, they will have
different managers and cause conflicts. However, most of
those cases can be solved with the current priority,
which we have explicitly documented in this commit.