Guard against invalid input for `keys`
I think it'd be nice for `Map.drop/2` and `Map.take/2` to guard against invalid inputs for the `keys` argument. I recognize that this isn't necessary, since internally the call to `Enum.to_list/1` raises for invalid inputs in these scenarios: users currently receive a `Protocol.UndefinedError` that reads `protocol Enumerable not implemented for...`.

This makes sense given the erroneous call to `Enum.to_list/1` on a non-enumerable. However this produces an error that reveals more about the implementation details of the `take/2` and `drop/2` functions than it reveals about the more immediately erroneous unsupported function call. I believe that a `FunctionClauseError` or an `ArgumentError` is perhaps more useful and instructive.