Keep one path per app for consolidation
The previous version shared the protocol consolidation
path between apps in the same umbrella. However, we would
only validate the protocol manifest if any artifact was
compiled. This caused issues in some situations such as:

  1. Compile the whole umbrella (which would consolidate)
  2. Compile an app that is a subset of the umbrella (which would consolidate)
  3. Run another subset in the umbrella (which would not consolidate)

Because the whole umbrella is compiled after step 1, any
further change wouldn't cause reconsolidation. This would
be any issue if any app was compiled with --force, as it
would force consolidation to a local snapshop which would
be reconsolidated when changing apps.

One solution to this problem would be to store in the
manifest which app owns the manifest and check it every
time. However, this approach would add complexity and
force reconsolidation every time we changed apps.

The approach taken in this commit is to have a consolidation
path per app. This may be a backwards incompatible change
if some were assuming the consolidation path is always the
same. We recommend developers to rely on
Mix.Project.consolidation_path which exists since v1.3.

Closes #5987