Streamline child specs (#6189)
This commit introduces a child_spec/1 function that is responsible
for controlling how a child process runs in a supervision tree.

This proposal also syncs Elixir with OTP by allowing child specs
to be given by maps. Supervisor.start_link/2 has been changed
accordingly and Supervisor.init/2 was added as a replacement for
the old Supervisor.Spec functionality.

The child_spec/1 function is automatically defined when using
GenServer, Supervisor, Agent and Task. A child_spec/1 was
also added to Elixir services such as Agent, Task, Task.Supervisor
and Registry.