Optimize String.replace_* functions
* Skip unnecessary matches on the "rest" size
* Don't prepend/append empty strings

The second one is especially important - doing "" <> str, will force a copy of
the string. Checking, if the string is empty first, allows us to skip this copy.
Using replace functions with empty replacements is rather common - one example
is the String.trim* family of functions, but even in user code it's quite
frequent.
Removing excessive copy, means the speedup can be made arbitrarily large
depending on input. The slowdown for cases with non-empty strings couldn't be
reliably measured.