Use orelse instead of or when expanding in operator (#6314)
orelse may lead to better code in many occasions - Since we're comparing
to literals, there's no semantic difference.

This also means, the compiler is able to figure out more "will always match"
checks. For example in a situation like the following a warning will be emitted.
This pull request also fixes such warnings in the test suite.

    if true do
      IO.puts("true")
    end

One problem with using `orelse` is that it will short-circuit and not evaluate
some side-effecting things, for example:

    1 in [1, send(self(), :foo)]

To remedy this, we'll traverse the right side of `in` and extract any things
that are not simple literals to variables, so that the above example expands
to something like the following:

    var0 = send(seld(), :foo)
    var = 1
    var === 1 or var === var0

This guarantees all side-effects are performed and we still retain benefits of
the short-circuiting comparisons in the end.