Add mix test --slowest reporting to ExUnit
This adds a `--slowest N` CLI flag to `mix test`, which will print
timing information for the slowest test times. To ensure that reported
times are accurate enabling slow reporting will automatically set the
suite in `--trace` mode.

The `slowest` setting defaults to `0`, disabling it for normal test
runs. Setting it to any positive integer will report a cumalitve slow
test time, slow tests as a percentage of total time, and display
individual timings for each of the slowest tests.

Closes #6334