Allow (un)registering other processes via Registry
Currently, Elixir only allows you to register keys/values to the current
process. However, there are scenarios where the ability to register
another process to some topic or metadata is very useful. This change
implements `register/4` and `unregister/3`, both of which take a new
argument, which is the pid of the process which is involved in the
registration. The existing `register/3` and `unregister/2` functions now
delegate their work to these new implementations, passing `self()` as
the pid.