Add System.get_env!/1, which raises if the env var isn't set
There are times when forgetting to set an environment variable could
cause an app to silently misbehave. For example, if the environment
variable contains the salt value for hashing passwords and it's silently
converted to an empty string, the passwords will be unsalted.

In such situations, it's useful to explicitly raise an error about the
missing environment variable.