Raise exception when passing negative delta to `assert_in_delta`
I agreed with @josevalim's idea of raising an exception here rather
than coercing the negative delta, so I implemented that. As a user, I
would find that coercion to be unexpected behavior, and I personally
felt the value of being explicit here outweighed the convenience of the
coercion. I did, however, put in the exception message a helpful hint
in the case that the user did want the absolute value of that negative
delta instead.

Fixes #6439