Add option :return_errors to ParallelCompiler and ParallelRequire
This is motivated by the goal of allowing Mix compilers to return their diagnostic information in a standardized format. Before this change, `ParallelCompiler` calls `exit({:shutdown, 1})` upon encountering a compilation error, and there's no way to programatically retrieve the errors. This change allows `ParallelCompiler` and `ParallelRequire` to return warnings and errors if the option `:return_errors` is passed. Return values is `{:ok, modules, warnings} | {:error, errors}` when this option is passed. If the option is not passed, the return value is the same as before, and it will exit the process when encountering an error.

There seemed to be a lot of duplication between the code in `ParallelRequire` and `ParallelCompiler`, so to avoid having to duplicate these changes in both places, I added an option `:require` to `ParallelCompiler` to load files via `Code.require_files/1`. Now, `ParallelRequire` simply calls `ParallelCompiler` with this option, reducing code duplication.