Remove incorrect wording regarding test cases
Previously, we would use `case` to refer to the whole module
but a `case` is actually a *single test*. We have fixed this
by introducing a new `ExUnit.TestModule` struct and by adding
a new attribute called `:module` to the relevant places.

The changes were kept backwards compatible but we do recommend
developers to:

  * Access the `:module` field instead of `:case` field
  * Replace the `:case_started` and `:case_finished` callbacks
    in their formatters by `:module_started` and `:module_finished`

The previous callbacks will continue to receive `ExUnit.TestCase`
structs but the new callbacks will receive `ExUnit.TestModule`.

Unfortunately those field names and callbacks cannot be deprecated,
so they will be one of the backwards incompatible changes required
for Elixir 2.0.