Add Inspect.Algebra.container_doc/6
This function replaces surround_many as it attempts to fit
as many algebra documents on the same line if and only if
all documents are simple.

A simple document is a document made only of strings,
without breaks. This allows us to print lists of complex
structures with one element per line but lists of integers
or lists of atoms still try to fit as much as possible in a
single-line.