Update process.ex (#6650)
Improved the error message: the "or" conjuction isn't used that way in English (as it would be in French, for example).

Didn't find a way to keep the list of possibilities as a separate sentence: couldn't lead with "Either" as it is typically used to differentiate between only 2 options. Finally, leaving the "Oxford comma" properly indicates that the issues is "one of" the presented possibilities.