Stop using keyword lists in IEx.Info
We were returning info from the protocol as lists of "{atom, string}"
pairs with atoms being "titles" of info (such as "Reference modules").
This is a bit weird and it definitely becomes annoying with the new code
formatter because it formats things like ["Description": ...] as
[Description: ...] which looks weird.

This commit turns all atom keys into strings so that now info is a list
of "{string, string}" tuples.