Files with only comments are no longer always stale
At the beginning of compilation, initial `:source` records of the files that
needs to be compiled are added to an Agent process. These records are then
updated when the files are compiled and then written to the compile manifest.

The `:size` field in the initial record has a default of 0, and since files
without any code (they only have comments or newlines) are not compiled, this
is never updated with the actual size of the file before the compile manifest
is written to disk.

These files are always seen as stale because their size on disk does not match
the compile manifest.

This is fixed by adding the actual size of the file to the initial record.

Closes #7066