Add yellow asterisk in output for skipped tests
This will give some feedback to the user when they have skipped tests
in their suite instead of showing everything as all green.

There weren't any tests for this module, so I haven't added one here.
However, I'd be happy to go through and add some more tests to cover
this module if it would be helpful!