Add typespec for Float.ratio/1
While going through and adding the `@since` attributes, I notice that
this function was missing a typespec, so I figured I'd add it.

I went with `{pos_integer, pos_integer} | {neg_integer, pos_integer}`
as the return type instead of `{integer, pos_integer}` to make it clear
that we would never have a numerator of 0.