Replace `skipped` with `excluded`
This introduces the concept of an "excluded" test. This test is usually
excluded at the command line or in some other sort of configuration
option, but is _not_ excluded by using the `:skip` tag.

Right now that functionality to temporarily skip a test is disabled,
but in the coming commits we're going to re-enable that.