Update doc for DynamicSupervisor.start_child
The previous example made it seem like the arguments to `def start_child`: `foo, bar, baz` where being dropped by the `spec = MyWorker`. I *think* this would be the correct way to delegate the module and arguments to `DynamicSupervisor.start_child` in the implied case that MyWorker is using the new child specs. If not, I'd be curious about the preferred approach.