Optimize ExUnit manifest to store only failures
Since that is the only information we care about, we do not
have to store every test with its status. This allows us to
vastly simplify the manifest, as we no longer have to use an
`:entry` record. Reading and writing the manifest should be
faster now, since the manifest tends toward being empty.

Since we are only storing failures, I have renamed the manifest
(and the `:manifest_file` option and file path) to indicate that
it deals only with failures.

Signed-off-by: José Valim <jose.valim@plataformatec.com.br>