Remove invalid UTF-8 code from assertion failures (#7449)
* Basics of new test for invalid UTF8 strings added

Since there are some tests that need to reference specific line
numbers, I wanted to add a commit with working code but those line
numbers all updated before I added the part that makes ExUnit blow up
(removing the escaping on the invalid UTF8 string).

* Remove invalid UTF-8 code from assertion failures

We were having issues with trying to write invalid UTF-8 sequences to
the console in the CLI formatter because if it was possible, we would
try and print the code that was failing in the failure message. This is
totally helpful, but we unfortunately can't print invalid UTF-8, so if
the code failed because it was invalid UTF-8, we can't show the user
that code.

We do still get the sequence that failed, and the line number that it
was on, so I'm hoping that this is still going to be enough for folks
to fix whatever issues they have.

Resolves #7434

* Update implementation to move remove of invalid UTF-8 earlier on

This version of the fix for this problem removes the `expr` field from
the `ExUnit.AssertionError` struct if that code isn't valid UTF-8.