Allow exclude patterns .formatter.exs
With wildcard support, the formatter configuration makes it very
easy to include whole directories full of files for formatting.
However, excluding a single file is not as trivial. While this is
not a regular use-case, there are definitely situations where it
is very useful.

This adds support for negating include patterns in the config
analog to how .gitignore works. The paths are resolved the same
way normal include paths are resolved and dropped from the file
list before passing it to the code formatter.

This allows for configurations like the following:
inputs: ["lib/**/*.{ex,exs}", "!lib/static.{ex,exs}"]