added info on child spec formats
Referring to the `Supervisor` docs enables the user to determine what the various child spec shapes relate to, and how they can be used to specify specs for dynamic children. In particular, it clarifies that the user doesn't need to provide a full child spec map, but can instead use the module or tuple forms.