Move `Code.string_to_quoted` error logic up
As discussed in #7270, it would help if we moved up the error logic out of Erlang. So, instead of raising an exception, we'll return an error tuple.
Currently, this will only fix the Hex error use case, so does not completely resolve this issue. But, if this is the route we want to take, I can work on the other usecases.