Modified `kernel.ex` intro
Another modification would be to reorder the list of functions in the nav bar into 4 sections: syntactic sugar, macros, guard checks, and other. Visually, just insert a line break in between each group.

I can implement this revision (and review the page in its entirety), if you agree with the direction of this pull request.